Exo 2 Data Done, Understanding Shiny
I did the Data aspect of Exo 2, so we have decent data that works with what we want to do. To use this data I found 2 ways, saving it in directory, or in environment. So now we have to see which works best with the Shiny App (probably directory but idk how)